scanner: internal handling of token space-separation
The scanner is walking a fine line dealing with whitespace, since tokens
are wsp-separated and thus it is an implicit part of the grammar.
In order not to make the parser more complex, whitespace token
separation is internally handled in the scanner.

This commit also fixes invalidly accepted keywords beginning with valid
keyword sequence.